Implement CustomerOrderController. Not tested.
This is almost line for line a copy-paste of the ResupplyOrderController.
It has significantly more commenting and a little more forethought into
exactly what is doing what. I basically went through and copied line
by line so as to better understand what is going on and how everything is
working. While developing ResupplyOrderController, I really had no idea
what I was doing or how things were working but have a much better understanding
at this point.